fixed: profile-local save games sometimes didn't work
this problem occurred when the ...Prefs.ini file also contained a SLocalSavePath field which, afaict was only common on Fallout NV - for whatever reason.
Worse yet: For some reason, if one sets the field in prefs but the regular ini contains the default, the game would still use the default. And the NV launcher resets the parameter if started so that's a pain....