fixed: exception trying to import savegames to same profile
The way to reproduce this is: go to savegame page, start the import process, select a profile that has savegames, then switch to that profile and go back to the savegames page.
Now the list still contains the same savegames, importing one would try to copy the save onto itself.

fixes Nexus-Mods/Vortex#2056